Added: API Tales of Woe and Woah
This is basically a combination of the parts I enjoyed of API Pain Points and What They Should Have Told You... without the boring THIS IS WHAT REST AND RPC MEAN which I have written up as an article instead. It was so boring to talk about I was rushing through it and it was fucking with my calm. This new one should be a chuckle, back to the format of the old API Pain Points stand-up style of talk.